requirements.txt
django-admin
django==4.0.6
django-ckeditor==6.5.1
django-js-asset==2.0.0
Pillow==9.2.0
tzdata==2022.1
mysqlclient==2.2.4
psycopg2-binary==2.9.9
requests==2.31.0